+++ src/main/java/tn/esprit/spring/services/IEmployeService.java
@@ -16,7 +16,7 @@ public interface IEmployeService {
 	public void mettreAjourEmailByEmployeId(String email, int employeId);
 	public void affecterEmployeADepartement(int employeId, int depId);
 	public void desaffecterEmployeDuDepartement(int employeId, int depId);
-	public int ajouterContrat(Contrat contrat);
+	public Long ajouterContrat(Contrat contrat);
 	public void affecterContratAEmploye(int contratId, int employeId);
 	public String getEmployePrenomById(int employeId);
 	public void deleteEmployeById(int employeId);
